Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

JSOSolvers.LBFGSSolver

JSOSolvers.TronSolver

JSOSolvers.TrunkSolver

JSOSolvers.cauchy

JSOSolvers.cauchy_ls

JSOSolvers.compute_As_slope_qs!

JSOSolvers.lbfgs

JSOSolvers.projected_gauss_newton!

JSOSolvers.projected_line_search!

JSOSolvers.projected_line_search_ls!

JSOSolvers.projected_newton!

JSOSolvers.tron

JSOSolvers.tron

JSOSolvers.trunk

JSOSolvers.trunk

Krylov.solve!

​

JSOSolvers.LBFGSSolver — Type

lbfgs(nlp)

solver = LBFGSSolver(nlp)
solve!(solver, nlp)

An implementation of a limited memory BFGS line-search method for unconstrained minimization.

JSOSolvers.TrunkSolver — Type

trunk(nlp)

solver = TrunkSolver(nlp)
output = solve!(solver, nlp)

A trust-region solver for unconstrained optimization using exact second derivatives.

This implementation follows the description given in [1]. The main algorithm follows the basic trust-region method described in Section 6. The backtracking linesearch follows Section 10.3.2. The nonmonotone strategy follows Section 10.1.3, Algorithm 10.1.2.

[1] A. R. Conn, N. I. M. Gould, and Ph. L. Toint, Trust-Region Methods, volume 1 of MPS/SIAM Series on Optimization. SIAM, Philadelphia, USA, 2000. DOI: 10.1137/1.9780898719857.

JSOSolvers.lbfgs — Method

lbfgs(nlp)

solver = LBFGSSolver(nlp)
solve!(solver, nlp)

An implementation of a limited memory BFGS line-search method for unconstrained minimization.

JSOSolvers.tron — Method

tron(nls)

A pure Julia implementation of a trust-region solver for bound-constrained nonlinear least squares optimization:

min ½‖F(x)‖²    s.t.    ℓ ≦ x ≦ u

This is an adaptation of the TRON method described in

Chih-Jen Lin and Jorge J. Moré, Newton's Method for Large Bound-Constrained Optimization Problems, SIAM J. Optim., 9(4), 1100–1127, 1999.

JSOSolvers.tron — Method

tron(nlp)

solver = TronSolver(nlp)
output = solve!(solver, nlp)

A pure Julia implementation of a trust-region solver for bound-constrained optimization:

min f(x)    s.t.    ℓ ≦ x ≦ u

TRON is described in

Chih-Jen Lin and Jorge J. Moré, Newton's Method for Large Bound-Constrained Optimization Problems, SIAM J. Optim., 9(4), 1100–1127, 1999.

JSOSolvers.trunk — Method

trunk(nls)

A trust-region solver for nonlinear least squares.

This implementation follows the description given in [1]. The main algorithm follows the basic trust-region method described in Section 6. The backtracking linesearch follows Section 10.3.2. The nonmonotone strategy follows Section 10.1.3, Algorithm 10.1.2.

[1] A. R. Conn, N. I. M. Gould, and Ph. L. Toint, Trust-Region Methods, volume 1 of MPS/SIAM Series on Optimization. SIAM, Philadelphia, USA, 2000. DOI: 10.1137/1.9780898719857.

JSOSolvers.trunk — Method

trunk(nlp)

solver = TrunkSolver(nlp)
output = solve!(solver, nlp)

A trust-region solver for unconstrained optimization using exact second derivatives.

This implementation follows the description given in [1]. The main algorithm follows the basic trust-region method described in Section 6. The backtracking linesearch follows Section 10.3.2. The nonmonotone strategy follows Section 10.1.3, Algorithm 10.1.2.

[1] A. R. Conn, N. I. M. Gould, and Ph. L. Toint, Trust-Region Methods, volume 1 of MPS/SIAM Series on Optimization. SIAM, Philadelphia, USA, 2000. DOI: 10.1137/1.9780898719857.

Krylov.solve! — Function

solve!(solver, nlp)

Solve nlp using solver.

JSOSolvers.TronSolver — Type

tron(nlp)

solver = TronSolver(nlp)
output = solve!(solver, nlp)

A pure Julia implementation of a trust-region solver for bound-constrained optimization:

min f(x)    s.t.    ℓ ≦ x ≦ u

TRON is described in

Chih-Jen Lin and Jorge J. Moré, Newton's Method for Large Bound-Constrained Optimization Problems, SIAM J. Optim., 9(4), 1100–1127, 1999.

JSOSolvers.cauchy — Method

α, s = cauchy(x, H, g, Δ, ℓ, u; μ₀ = 1e-2, μ₁ = 1.0, σ=10.0)

Computes a Cauchy step s = P(x - α g) - x for

min  q(s) = ¹/₂sᵀHs + gᵀs     s.t.    ‖s‖ ≦ μ₁Δ,  ℓ ≦ x + s ≦ u,

with the sufficient decrease condition

q(s) ≦ μ₀sᵀg.

JSOSolvers.cauchy_ls — Method

α, s = cauchy_ls(x, A, Fx, g, Δ, ℓ, u; μ₀ = 1e-2, μ₁ = 1.0, σ=10.0)

Computes a Cauchy step s = P(x - α g) - x for

min  q(s) = ½‖As + Fx‖² - ½‖Fx‖²     s.t.    ‖s‖ ≦ μ₁Δ,  ℓ ≦ x + s ≦ u,

with the sufficient decrease condition

q(s) ≦ μ₀gᵀs,

where g = AᵀFx.

JSOSolvers.compute_As_slope_qs! — Method

slope, qs = compute_As_slope_qs!(As, A, s, Fx)

Compute slope = dot(As, Fx) and qs = dot(As, As) / 2 + slope. Use As to store A * s.

JSOSolvers.projected_gauss_newton! — Method

projected_gauss_newton!(x, A, Fx, Δ, gctol, s, max_cgiter, ℓ, u)

Compute an approximate solution d for

min q(d) = ½‖Ad + Fx‖² - ½‖Fx‖²     s.t.    ℓ ≦ x + d ≦ u,  ‖d‖ ≦ Δ

starting from s. The steps are computed using the conjugate gradient method projected on the active bounds.

JSOSolvers.projected_line_search! — Method

s = projected_line_search!(x, H, g, d, ℓ, u; μ₀ = 1e-2)

Performs a projected line search, searching for a step size t such that

0.5sᵀHs + sᵀg ≦ μ₀sᵀg,

where s = P(x + t * d) - x, while remaining on the same face as x + d. Backtracking is performed from t = 1.0. x is updated in place.

JSOSolvers.projected_line_search_ls! — Method

s = projected_line_search_ls!(x, A, g, d, ℓ, u; μ₀ = 1e-2)

Performs a projected line search, searching for a step size t such that

½‖As + Fx‖² ≤ ½‖Fx‖² + μ₀FxᵀAs

where s = P(x + t * d) - x, while remaining on the same face as x + d. Backtracking is performed from t = 1.0. x is updated in place.

JSOSolvers.projected_newton! — Method

projected_newton!(x, H, g, Δ, gctol, s, max_cgiter, ℓ, u)

Compute an approximate solution d for

min q(d) = ¹/₂dᵀHs + dᵀg s.t. ℓ ≦ x + d ≦ u, ‖d‖ ≦ Δ

starting from s. The steps are computed using the conjugate gradient method projected on the active bounds.