Solvers

Solver list

lbfgs

tron

trunk

Problem type	Solvers

Unconstrained NLP	lbfgs, tron, trunk

Unconstrained NLS	trunk, tron

Bound-constrained NLP	tron

Bound-constrained NLS	tron

Solver list

JSOSolvers.lbfgs — Function

lbfgs(nlp)

solver = LBFGSSolver(nlp)
solve!(solver, nlp)

An implementation of a limited memory BFGS line-search method for unconstrained minimization.

JSOSolvers.tron — Function

tron(nlp)

solver = TronSolver(nlp)
output = solve!(solver, nlp)

A pure Julia implementation of a trust-region solver for bound-constrained optimization:

min f(x)    s.t.    ℓ ≦ x ≦ u

TRON is described in

Chih-Jen Lin and Jorge J. Moré, Newton's Method for Large Bound-Constrained Optimization Problems, SIAM J. Optim., 9(4), 1100–1127, 1999.

tron(nls)

A pure Julia implementation of a trust-region solver for bound-constrained nonlinear least squares optimization:

min ½‖F(x)‖²    s.t.    ℓ ≦ x ≦ u

This is an adaptation of the TRON method described in

Chih-Jen Lin and Jorge J. Moré, Newton's Method for Large Bound-Constrained Optimization Problems, SIAM J. Optim., 9(4), 1100–1127, 1999.

JSOSolvers.trunk — Function

trunk(nlp; kwargs...)

A trust-region solver for unconstrained optimization using exact second derivatives.

For an advanced usage, one can first define an TrunkSolver preallocating the memory used in the algorithm and then call solve!.

solver = TrunkSolver(nlp, subsolver_type::Type{<:KrylovSolver} = CgSolver)
solve!(solver, nlp; kwargs...)

Arguments

nlp::AbstractNLPModel represents the model solved, see NLPModels.jl.

The keyword arguments may include

subsolver_logger::AbstractLogger = NullLogger(): subproblem's logger.

x = nlp.meta.x0: the initial guess.

atol::T = √eps(T): absolute tolerance.

rtol::T = √eps(T): relative tolerance, the algorithm stops when ||∇f(xᵏ)|| ≤ atol + rtol * ||∇f(x⁰)||.

max_eval::Int = -1: maximum number of objective function evaluations.

max_time::Float64 = 30.0: maximum time limit.

bk_max::Int = 10: algorithm's parameter.

monotone::Bool = true: algorithm's parameter.

nm_itmax::Int = 25: algorithm's parameter.

verbose::Int = 0: If > 0, display interation information every verbose iteration.

verbose_subsolver::Int = 0: If > 0, display interation information every verbose_subsolver iteration of the subsolver.

Output

The returned value is a GenericExecutionStats, see SolverCore.jl.

References

This implementation follows the description given in

A. R. Conn, N. I. M. Gould, and Ph. L. Toint,
Trust-Region Methods, volume 1 of MPS/SIAM Series on Optimization.
SIAM, Philadelphia, USA, 2000.
DOI: 10.1137/1.9780898719857

The main algorithm follows the basic trust-region method described in Section 6. The backtracking linesearch follows Section 10.3.2. The nonmonotone strategy follows Section 10.1.3, Algorithm 10.1.2.

Examples

using JSOSolvers, ADNLPModels
nlp = ADNLPModel(x -> sum(x.^2), ones(3))
stats = trunk(nlp)

using JSOSolvers, ADNLPModels
nlp = ADNLPModel(x -> sum(x.^2), ones(3))
solver = TrunkSolver(nlp)
stats = solve!(solver, nlp)

trunk(nls)

A trust-region solver for nonlinear least squares.

This implementation follows the description given in [1]. The main algorithm follows the basic trust-region method described in Section 6. The backtracking linesearch follows Section 10.3.2. The nonmonotone strategy follows Section 10.1.3, Algorithm 10.1.2.

[1] A. R. Conn, N. I. M. Gould, and Ph. L. Toint, Trust-Region Methods, volume 1 of MPS/SIAM Series on Optimization. SIAM, Philadelphia, USA, 2000. DOI: 10.1137/1.9780898719857.