Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

JSOSolvers.LBFGSSolver

JSOSolvers.R2Solver

JSOSolvers.TronSolver

JSOSolvers.TronSolverNLS

JSOSolvers.TrunkSolver

JSOSolvers.TrunkSolverNLS

JSOSolvers.R2

JSOSolvers.cauchy!

JSOSolvers.cauchy_ls

JSOSolvers.compute_As_slope_qs!

JSOSolvers.lbfgs

JSOSolvers.projected_gauss_newton!

JSOSolvers.projected_line_search!

JSOSolvers.projected_line_search_ls!

JSOSolvers.projected_newton!

JSOSolvers.tron

JSOSolvers.tron

JSOSolvers.trunk

JSOSolvers.trunk

​

JSOSolvers.LBFGSSolver — Type

lbfgs(nlp; kwargs...)

An implementation of a limited memory BFGS line-search method for unconstrained minimization.

For advanced usage, first define a LBFGSSolver to preallocate the memory used in the algorithm, and then call solve!.

solver = LBFGSSolver(nlp; mem::Int = 5)
solve!(solver, nlp; kwargs...)

Arguments

nlp::AbstractNLPModel{T, V} represents the model to solve, see NLPModels.jl.

The keyword arguments may include

x::V = nlp.meta.x0: the initial guess.

mem::Int = 5: memory parameter of the lbfgs algorithm.

atol::T = √eps(T): absolute tolerance.

rtol::T = √eps(T): relative tolerance, the algorithm stops when ‖∇f(xᵏ)‖ ≤ atol + rtol * ‖∇f(x⁰)‖.

max_eval::Int = -1: maximum number of objective function evaluations.

max_time::Float64 = 30.0: maximum time limit in seconds.

max_iter::Int = typemax(Int): maximum number of iterations.

τ₁::T = T(0.9999): slope factor in the Wolfe condition when performing the line search.

bk_max:: Int = 25: maximum number of backtracks when performing the line search.

verbose::Int = 0: if > 0, display iteration details every verbose iteration.

verbose_subsolver::Int = 0: if > 0, display iteration information every verbose_subsolver iteration of the subsolver.

Output

The returned value is a GenericExecutionStats, see SolverCore.jl.

Callback

The callback is called at each iteration. The expected signature of the callback is callback(nlp, solver, stats), and its output is ignored. Changing any of the input arguments will affect the subsequent iterations. In particular, setting stats.status = :user will stop the algorithm. All relevant information should be available in nlp and solver. Notably, you can access, and modify, the following:

solver.x: current iterate;

solver.gx: current gradient;

stats: structure holding the output of the algorithm (GenericExecutionStats), which contains, among other things:

stats.dual_feas: norm of current gradient;

stats.iter: current iteration counter;

stats.objective: current objective function value;

stats.status: current status of the algorithm. Should be :unknown unless the algorithm has found a stopping criteria. Changing this to anything will stop the algorithm, but you should use :user to properly indicate the intention.

stats.elapsed_time: elapsed time in seconds.

Examples

using JSOSolvers, ADNLPModels
nlp = ADNLPModel(x -> sum(x.^2), ones(3));
stats = lbfgs(nlp)

# output

"Execution stats: first-order stationary"

using JSOSolvers, ADNLPModels
nlp = ADNLPModel(x -> sum(x.^2), ones(3));
solver = LBFGSSolver(nlp; mem = 5);
stats = solve!(solver, nlp)

# output

"Execution stats: first-order stationary"

JSOSolvers.R2Solver — Type

R2(nlp; kwargs...)

A first-order quadratic regularization method for unconstrained optimization.

For advanced usage, first define a R2Solver to preallocate the memory used in the algorithm, and then call solve!:

solver = R2Solver(nlp)
solve!(solver, nlp; kwargs...)

Arguments

nlp::AbstractNLPModel{T, V} is the model to solve, see NLPModels.jl.

Keyword arguments

x::V = nlp.meta.x0: the initial guess.

atol::T = √eps(T): absolute tolerance.

rtol::T = √eps(T): relative tolerance: algorithm stops when ‖∇f(xᵏ)‖ ≤ atol + rtol * ‖∇f(x⁰)‖.

η1 = eps(T)^(1/4), η2 = T(0.95): step acceptance parameters.

γ1 = T(1/2), γ2 = 1/γ1: regularization update parameters.

σmin = eps(T): step parameter for R2 algorithm.

max_eval::Int = -1: maximum number of evaluation of the objective function.

max_time::Float64 = 30.0: maximum time limit in seconds.

max_iter::Int = typemax(Int): maximum number of iterations.

β = T(0) ∈ [0,1] is the constant in the momentum term. If β == 0, R2 does not use momentum.

verbose::Int = 0: if > 0, display iteration details every verbose iteration.

Output

The value returned is a GenericExecutionStats, see SolverCore.jl.

Callback

The callback is called at each iteration. The expected signature of the callback is callback(nlp, solver, stats), and its output is ignored. Changing any of the input arguments will affect the subsequent iterations. In particular, setting stats.status = :user will stop the algorithm. All relevant information should be available in nlp and solver. Notably, you can access, and modify, the following:

solver.x: current iterate;

solver.gx: current gradient;

stats: structure holding the output of the algorithm (GenericExecutionStats), which contains, among other things:

stats.dual_feas: norm of current gradient;

stats.iter: current iteration counter;

stats.objective: current objective function value;

stats.status: current status of the algorithm. Should be :unknown unless the algorithm has attained a stopping criterion. Changing this to anything will stop the algorithm, but you should use :user to properly indicate the intention.

stats.elapsed_time: elapsed time in seconds.

Examples

using JSOSolvers, ADNLPModels
nlp = ADNLPModel(x -> sum(x.^2), ones(3))
stats = R2(nlp)

# output

"Execution stats: first-order stationary"

using JSOSolvers, ADNLPModels
nlp = ADNLPModel(x -> sum(x.^2), ones(3))
solver = R2Solver(nlp);
stats = solve!(solver, nlp)

# output

"Execution stats: first-order stationary"

JSOSolvers.TronSolver — Type

tron(nlp; kwargs...)

A pure Julia implementation of a trust-region solver for bound-constrained optimization:

    min f(x)    s.t.    ℓ ≦ x ≦ u

For advanced usage, first define a TronSolver to preallocate the memory used in the algorithm, and then call solve!:

solver = TronSolver(nlp; kwargs...)
solve!(solver, nlp; kwargs...)

Arguments

nlp::AbstractNLPModel{T, V} represents the model to solve, see NLPModels.jl.

The keyword arguments may include

x::V = nlp.meta.x0: the initial guess.

μ₀::T = T(1e-2): algorithm parameter in (0, 0.5).

μ₁::T = one(T): algorithm parameter in (0, +∞).

σ::T = T(10): algorithm parameter in (1, +∞).

max_eval::Int = -1: maximum number of objective function evaluations.

max_time::Float64 = 30.0: maximum time limit in seconds.

max_iter::Int = typemax(Int): maximum number of iterations.

max_cgiter::Int = 50: subproblem's iteration limit.

use_only_objgrad::Bool = false: If true, the algorithm uses only the function objgrad instead of obj and grad.

cgtol::T = T(0.1): subproblem tolerance.

atol::T = √eps(T): absolute tolerance.

rtol::T = √eps(T): relative tolerance, the algorithm stops when ‖∇f(xᵏ)‖ ≤ atol + rtol * ‖∇f(x⁰)‖.

verbose::Int = 0: if > 0, display iteration details every verbose iteration.

The keyword arguments of TronSolver are passed to the TRONTrustRegion constructor.

Output

The value returned is a GenericExecutionStats, see SolverCore.jl.

Callback

The callback is called at each iteration. The expected signature of the callback is callback(nlp, solver, stats), and its output is ignored. Changing any of the input arguments will affect the subsequent iterations. In particular, setting stats.status = :user will stop the algorithm. All relevant information should be available in nlp and solver. Notably, you can access, and modify, the following:

solver.x: current iterate;

solver.gx: current gradient;

stats: structure holding the output of the algorithm (GenericExecutionStats), which contains, among other things:

stats.dual_feas: norm of current gradient;

stats.iter: current iteration counter;

stats.objective: current objective function value;

stats.status: current status of the algorithm. Should be :unknown unless the algorithm has attained a stopping criterion. Changing this to anything will stop the algorithm, but you should use :user to properly indicate the intention.

stats.elapsed_time: elapsed time in seconds.

References

TRON is described in

Chih-Jen Lin and Jorge J. Moré, *Newton's Method for Large Bound-Constrained
Optimization Problems*, SIAM J. Optim., 9(4), 1100–1127, 1999.
DOI: 10.1137/S1052623498345075

Examples

using JSOSolvers, ADNLPModels
nlp = ADNLPModel(x -> sum(x), ones(3), zeros(3), 2 * ones(3));
stats = tron(nlp)

using JSOSolvers, ADNLPModels
nlp = ADNLPModel(x -> sum(x), ones(3), zeros(3), 2 * ones(3));
solver = TronSolver(nlp);
stats = solve!(solver, nlp)

JSOSolvers.TronSolverNLS — Type

tron(nls; kwargs...)

A pure Julia implementation of a trust-region solver for bound-constrained nonlinear least-squares problems:

min ½‖F(x)‖²    s.t.    ℓ ≦ x ≦ u

For advanced usage, first define a TronSolverNLS to preallocate the memory used in the algorithm, and then call solve!: solver = TronSolverNLS(nls; kwargs...) solve!(solver, nls; kwargs...)

Arguments

nls::AbstractNLSModel{T, V} represents the model to solve, see NLPModels.jl.

The keyword arguments may include

subsolver_logger::AbstractLogger = NullLogger(): subproblem's logger.

x::V = nlp.meta.x0: the initial guess.

subsolver::Symbol = :lsmr: Krylov.jl method used as subproblem solver, see JSOSolvers.tronls_allowed_subsolvers for a list.

μ₀::T = T(1e-2): algorithm parameter in (0, 0.5).

μ₁::T = one(T): algorithm parameter in (0, +∞).

σ::T = T(10): algorithm parameter in (1, +∞).

max_eval::Int = -1: maximum number of objective function evaluations.

max_time::Float64 = 30.0: maximum time limit in seconds.

max_iter::Int = typemax(Int): maximum number of iterations.

max_cgiter::Int = 50: subproblem iteration limit.

cgtol::T = T(0.1): subproblem tolerance.

atol::T = √eps(T): absolute tolerance.

rtol::T = √eps(T): relative tolerance, the algorithm stops when ‖∇f(xᵏ)‖ ≤ atol + rtol * ‖∇f(x⁰)‖.

Fatol::T = √eps(T): absolute tolerance on the residual.

Frtol::T = eps(T): relative tolerance on the residual, the algorithm stops when ‖F(xᵏ)‖ ≤ Fatol + Frtol * ‖F(x⁰)‖.

verbose::Int = 0: if > 0, display iteration details every verbose iteration.

The keyword arguments of TronSolverNLS are passed to the TRONTrustRegion constructor.

Output

The value returned is a GenericExecutionStats, see SolverCore.jl.

Callback

The callback is called at each iteration. The expected signature of the callback is callback(nlp, solver, stats), and its output is ignored. Changing any of the input arguments will affect the subsequent iterations. In particular, setting stats.status = :user will stop the algorithm. All relevant information should be available in nlp and solver. Notably, you can access, and modify, the following:

solver.x: current iterate;

solver.gx: current gradient;

stats: structure holding the output of the algorithm (GenericExecutionStats), which contains, among other things:

stats.dual_feas: norm of current gradient;

stats.iter: current iteration counter;

stats.objective: current objective function value;

stats.status: current status of the algorithm. Should be :unknown unless the algorithm attained a stopping criterion. Changing this to anything will stop the algorithm, but you should use :user to properly indicate the intention.

stats.elapsed_time: elapsed time in seconds.

References

This is an adaptation for bound-constrained nonlinear least-squares problems of the TRON method described in

Chih-Jen Lin and Jorge J. Moré, *Newton's Method for Large Bound-Constrained
Optimization Problems*, SIAM J. Optim., 9(4), 1100–1127, 1999.
DOI: 10.1137/S1052623498345075

Examples

using JSOSolvers, ADNLPModels
F(x) = [x[1] - 1.0; 10 * (x[2] - x[1]^2)]
x0 = [-1.2; 1.0]
nls = ADNLSModel(F, x0, 2, zeros(2), 0.5 * ones(2))
stats = tron(nls)

using JSOSolvers, ADNLPModels
F(x) = [x[1] - 1.0; 10 * (x[2] - x[1]^2)]
x0 = [-1.2; 1.0]
nls = ADNLSModel(F, x0, 2, zeros(2), 0.5 * ones(2))
solver = TronSolverNLS(nls)
stats = solve!(solver, nls)

JSOSolvers.TrunkSolver — Type

trunk(nlp; kwargs...)

A trust-region solver for unconstrained optimization using exact second derivatives.

For advanced usage, first define a TrunkSolver to preallocate the memory used in the algorithm, and then call solve!:

solver = TrunkSolver(nlp, subsolver_type::Type{<:KrylovSolver} = CgSolver)
solve!(solver, nlp; kwargs...)

Arguments

nlp::AbstractNLPModel{T, V} represents the model to solve, see NLPModels.jl.

The keyword arguments may include

subsolver_logger::AbstractLogger = NullLogger(): subproblem's logger.

x::V = nlp.meta.x0: the initial guess.

atol::T = √eps(T): absolute tolerance.

rtol::T = √eps(T): relative tolerance, the algorithm stops when ‖∇f(xᵏ)‖ ≤ atol + rtol * ‖∇f(x⁰)‖.

max_eval::Int = -1: maximum number of objective function evaluations.

max_time::Float64 = 30.0: maximum time limit in seconds.

max_iter::Int = typemax(Int): maximum number of iterations.

bk_max::Int = 10: algorithm parameter.

monotone::Bool = true: algorithm parameter.

nm_itmax::Int = 25: algorithm parameter.

verbose::Int = 0: if > 0, display iteration information every verbose iteration.

verbose_subsolver::Int = 0: if > 0, display iteration information every verbose_subsolver iteration of the subsolver.

Output

The returned value is a GenericExecutionStats, see SolverCore.jl.

Callback

The callback is called at each iteration. The expected signature of the callback is callback(nlp, solver, stats), and its output is ignored. Changing any of the input arguments will affect the subsequent iterations. In particular, setting stats.status = :user will stop the algorithm. All relevant information should be available in nlp and solver. Notably, you can access, and modify, the following:

solver.x: current iterate;

solver.gx: current gradient;

stats: structure holding the output of the algorithm (GenericExecutionStats), which contains, among other things:

stats.dual_feas: norm of current gradient;

stats.iter: current iteration counter;

stats.objective: current objective function value;

stats.status: current status of the algorithm. Should be :unknown unless the algorithm attained a stopping criterion. Changing this to anything will stop the algorithm, but you should use :user to properly indicate the intention.

stats.elapsed_time: elapsed time in seconds.

References

This implementation follows the description given in

A. R. Conn, N. I. M. Gould, and Ph. L. Toint,
Trust-Region Methods, volume 1 of MPS/SIAM Series on Optimization.
SIAM, Philadelphia, USA, 2000.
DOI: 10.1137/1.9780898719857

The main algorithm follows the basic trust-region method described in Section 6. The backtracking linesearch follows Section 10.3.2. The nonmonotone strategy follows Section 10.1.3, Algorithm 10.1.2.

Examples

using JSOSolvers, ADNLPModels
nlp = ADNLPModel(x -> sum(x.^2), ones(3))
stats = trunk(nlp)

using JSOSolvers, ADNLPModels
nlp = ADNLPModel(x -> sum(x.^2), ones(3))
solver = TrunkSolver(nlp)
stats = solve!(solver, nlp)

JSOSolvers.TrunkSolverNLS — Type

trunk(nls; kwargs...)

A pure Julia implementation of a trust-region solver for nonlinear least-squares problems:

min ½‖F(x)‖²

For advanced usage, first define a TrunkSolverNLS to preallocate the memory used in the algorithm, and then call solve!:

solver = TrunkSolverNLS(nls, subsolver_type::Type{<:KrylovSolver} = LsmrSolver)
solve!(solver, nls; kwargs...)

Arguments

nls::AbstractNLSModel{T, V} represents the model to solve, see NLPModels.jl.

The keyword arguments may include

x::V = nlp.meta.x0: the initial guess.

atol::T = √eps(T): absolute tolerance.

rtol::T = √eps(T): relative tolerance, the algorithm stops when ‖∇f(xᵏ)‖ ≤ atol + rtol * ‖∇f(x⁰)‖.

Fatol::T = √eps(T): absolute tolerance on the residual.

Frtol::T = eps(T): relative tolerance on the residual, the algorithm stops when ‖F(xᵏ)‖ ≤ Fatol + Frtol * ‖F(x⁰)‖.

max_eval::Int = -1: maximum number of objective function evaluations.

max_time::Float64 = 30.0: maximum time limit in seconds.

max_iter::Int = typemax(Int): maximum number of iterations.

bk_max::Int = 10: algorithm parameter.

monotone::Bool = true: algorithm parameter.

nm_itmax::Int = 25: algorithm parameter.

verbose::Int = 0: if > 0, display iteration details every verbose iteration.

verbose_subsolver::Int = 0: if > 0, display iteration information every verbose_subsolver iteration of the subsolver.

See JSOSolvers.trunkls_allowed_subsolvers for a list of available KrylovSolver.

Output

The value returned is a GenericExecutionStats, see SolverCore.jl.

Callback

The callback is called at each iteration. The expected signature of the callback is callback(nlp, solver, stats), and its output is ignored. Changing any of the input arguments will affect the subsequent iterations. In particular, setting stats.status = :user will stop the algorithm. All relevant information should be available in nlp and solver. Notably, you can access, and modify, the following:

solver.x: current iterate;

solver.gx: current gradient;

stats: structure holding the output of the algorithm (GenericExecutionStats), which contains, among other things:

stats.dual_feas: norm of current gradient;

stats.iter: current iteration counter;

stats.objective: current objective function value;

stats.status: current status of the algorithm. Should be :unknown unless the algorithm attained a stopping criterion. Changing this to anything will stop the algorithm, but you should use :user to properly indicate the intention.

stats.elapsed_time: elapsed time in seconds.

References

This implementation follows the description given in

A. R. Conn, N. I. M. Gould, and Ph. L. Toint,
Trust-Region Methods, volume 1 of MPS/SIAM Series on Optimization.
SIAM, Philadelphia, USA, 2000.
DOI: 10.1137/1.9780898719857

The main algorithm follows the basic trust-region method described in Section 6. The backtracking linesearch follows Section 10.3.2. The nonmonotone strategy follows Section 10.1.3, Algorithm 10.1.2.

Examples

using JSOSolvers, ADNLPModels
F(x) = [x[1] - 1.0; 10 * (x[2] - x[1]^2)]
x0 = [-1.2; 1.0]
nls = ADNLSModel(F, x0, 2)
stats = trunk(nls)

using JSOSolvers, ADNLPModels
F(x) = [x[1] - 1.0; 10 * (x[2] - x[1]^2)]
x0 = [-1.2; 1.0]
nls = ADNLSModel(F, x0, 2)
solver = TrunkSolverNLS(nls)
stats = solve!(solver, nls)

JSOSolvers.R2 — Method

R2(nlp; kwargs...)

A first-order quadratic regularization method for unconstrained optimization.

For advanced usage, first define a R2Solver to preallocate the memory used in the algorithm, and then call solve!:

solver = R2Solver(nlp)
solve!(solver, nlp; kwargs...)

Arguments

nlp::AbstractNLPModel{T, V} is the model to solve, see NLPModels.jl.

Keyword arguments

x::V = nlp.meta.x0: the initial guess.

atol::T = √eps(T): absolute tolerance.

rtol::T = √eps(T): relative tolerance: algorithm stops when ‖∇f(xᵏ)‖ ≤ atol + rtol * ‖∇f(x⁰)‖.

η1 = eps(T)^(1/4), η2 = T(0.95): step acceptance parameters.

γ1 = T(1/2), γ2 = 1/γ1: regularization update parameters.

σmin = eps(T): step parameter for R2 algorithm.

max_eval::Int = -1: maximum number of evaluation of the objective function.

max_time::Float64 = 30.0: maximum time limit in seconds.

max_iter::Int = typemax(Int): maximum number of iterations.

β = T(0) ∈ [0,1] is the constant in the momentum term. If β == 0, R2 does not use momentum.

verbose::Int = 0: if > 0, display iteration details every verbose iteration.

Output

The value returned is a GenericExecutionStats, see SolverCore.jl.

Callback

The callback is called at each iteration. The expected signature of the callback is callback(nlp, solver, stats), and its output is ignored. Changing any of the input arguments will affect the subsequent iterations. In particular, setting stats.status = :user will stop the algorithm. All relevant information should be available in nlp and solver. Notably, you can access, and modify, the following:

solver.x: current iterate;

solver.gx: current gradient;

stats: structure holding the output of the algorithm (GenericExecutionStats), which contains, among other things:

stats.dual_feas: norm of current gradient;

stats.iter: current iteration counter;

stats.objective: current objective function value;

stats.status: current status of the algorithm. Should be :unknown unless the algorithm has attained a stopping criterion. Changing this to anything will stop the algorithm, but you should use :user to properly indicate the intention.

stats.elapsed_time: elapsed time in seconds.

Examples

using JSOSolvers, ADNLPModels
nlp = ADNLPModel(x -> sum(x.^2), ones(3))
stats = R2(nlp)

# output

"Execution stats: first-order stationary"

using JSOSolvers, ADNLPModels
nlp = ADNLPModel(x -> sum(x.^2), ones(3))
solver = R2Solver(nlp);
stats = solve!(solver, nlp)

# output

"Execution stats: first-order stationary"

JSOSolvers.cauchy! — Method

α, s = cauchy!(x, H, g, Δ, ℓ, u, s, Hs; μ₀ = 1e-2, μ₁ = 1.0, σ=10.0)

Computes a Cauchy step s = P(x - α g) - x for

min  q(s) = ¹/₂sᵀHs + gᵀs     s.t.    ‖s‖ ≦ μ₁Δ,  ℓ ≦ x + s ≦ u,

with the sufficient decrease condition

q(s) ≦ μ₀sᵀg.

JSOSolvers.cauchy_ls — Method

α, s = cauchy_ls(x, A, Fx, g, Δ, ℓ, u; μ₀ = 1e-2, μ₁ = 1.0, σ=10.0)

Computes a Cauchy step s = P(x - α g) - x for

min  q(s) = ½‖As + Fx‖² - ½‖Fx‖²     s.t.    ‖s‖ ≦ μ₁Δ,  ℓ ≦ x + s ≦ u,

with the sufficient decrease condition

q(s) ≦ μ₀gᵀs,

where g = AᵀFx.

JSOSolvers.compute_As_slope_qs! — Method

slope, qs = compute_As_slope_qs!(As, A, s, Fx)

Compute slope = dot(As, Fx) and qs = dot(As, As) / 2 + slope. Use As to store A * s.

JSOSolvers.lbfgs — Method

lbfgs(nlp; kwargs...)

An implementation of a limited memory BFGS line-search method for unconstrained minimization.

For advanced usage, first define a LBFGSSolver to preallocate the memory used in the algorithm, and then call solve!.

solver = LBFGSSolver(nlp; mem::Int = 5)
solve!(solver, nlp; kwargs...)

Arguments

nlp::AbstractNLPModel{T, V} represents the model to solve, see NLPModels.jl.

The keyword arguments may include

x::V = nlp.meta.x0: the initial guess.

mem::Int = 5: memory parameter of the lbfgs algorithm.

atol::T = √eps(T): absolute tolerance.

rtol::T = √eps(T): relative tolerance, the algorithm stops when ‖∇f(xᵏ)‖ ≤ atol + rtol * ‖∇f(x⁰)‖.

max_eval::Int = -1: maximum number of objective function evaluations.

max_time::Float64 = 30.0: maximum time limit in seconds.

max_iter::Int = typemax(Int): maximum number of iterations.

τ₁::T = T(0.9999): slope factor in the Wolfe condition when performing the line search.

bk_max:: Int = 25: maximum number of backtracks when performing the line search.

verbose::Int = 0: if > 0, display iteration details every verbose iteration.

verbose_subsolver::Int = 0: if > 0, display iteration information every verbose_subsolver iteration of the subsolver.

Output

The returned value is a GenericExecutionStats, see SolverCore.jl.

Callback

The callback is called at each iteration. The expected signature of the callback is callback(nlp, solver, stats), and its output is ignored. Changing any of the input arguments will affect the subsequent iterations. In particular, setting stats.status = :user will stop the algorithm. All relevant information should be available in nlp and solver. Notably, you can access, and modify, the following:

solver.x: current iterate;

solver.gx: current gradient;

stats: structure holding the output of the algorithm (GenericExecutionStats), which contains, among other things:

stats.dual_feas: norm of current gradient;

stats.iter: current iteration counter;

stats.objective: current objective function value;

stats.status: current status of the algorithm. Should be :unknown unless the algorithm has found a stopping criteria. Changing this to anything will stop the algorithm, but you should use :user to properly indicate the intention.

stats.elapsed_time: elapsed time in seconds.

Examples

using JSOSolvers, ADNLPModels
nlp = ADNLPModel(x -> sum(x.^2), ones(3));
stats = lbfgs(nlp)

# output

"Execution stats: first-order stationary"

using JSOSolvers, ADNLPModels
nlp = ADNLPModel(x -> sum(x.^2), ones(3));
solver = LBFGSSolver(nlp; mem = 5);
stats = solve!(solver, nlp)

# output

"Execution stats: first-order stationary"

JSOSolvers.projected_gauss_newton! — Method

projected_gauss_newton!(x, A, Fx, Δ, gctol, s, max_cgiter, ℓ, u)

Compute an approximate solution d for

min q(d) = ½‖Ad + Fx‖² - ½‖Fx‖²     s.t.    ℓ ≦ x + d ≦ u,  ‖d‖ ≦ Δ

starting from s. The steps are computed using the conjugate gradient method projected on the active bounds.

JSOSolvers.projected_line_search! — Method

s = projected_line_search!(x, H, g, d, ℓ, u, Hs; μ₀ = 1e-2)

Performs a projected line search, searching for a step size t such that

0.5sᵀHs + sᵀg ≦ μ₀sᵀg,

where s = P(x + t * d) - x, while remaining on the same face as x + d. Backtracking is performed from t = 1.0. x is updated in place.

JSOSolvers.projected_line_search_ls! — Method

s = projected_line_search_ls!(x, A, g, d, ℓ, u; μ₀ = 1e-2)

Performs a projected line search, searching for a step size t such that

½‖As + Fx‖² ≤ ½‖Fx‖² + μ₀FxᵀAs

where s = P(x + t * d) - x, while remaining on the same face as x + d. Backtracking is performed from t = 1.0. x is updated in place.

JSOSolvers.projected_newton! — Method

projected_newton!(solver, x, H, g, Δ, cgtol, ℓ, u, s, Hs; max_cgiter = 50)

Compute an approximate solution d for

min q(d) = ¹/₂dᵀHs + dᵀg s.t. ℓ ≦ x + d ≦ u, ‖d‖ ≦ Δ

starting from s. The steps are computed using the conjugate gradient method projected on the active bounds.

JSOSolvers.tron — Method

tron(nls; kwargs...)

A pure Julia implementation of a trust-region solver for bound-constrained nonlinear least-squares problems:

min ½‖F(x)‖²    s.t.    ℓ ≦ x ≦ u

For advanced usage, first define a TronSolverNLS to preallocate the memory used in the algorithm, and then call solve!: solver = TronSolverNLS(nls; kwargs...) solve!(solver, nls; kwargs...)

Arguments

nls::AbstractNLSModel{T, V} represents the model to solve, see NLPModels.jl.

The keyword arguments may include

subsolver_logger::AbstractLogger = NullLogger(): subproblem's logger.

x::V = nlp.meta.x0: the initial guess.

subsolver::Symbol = :lsmr: Krylov.jl method used as subproblem solver, see JSOSolvers.tronls_allowed_subsolvers for a list.

μ₀::T = T(1e-2): algorithm parameter in (0, 0.5).

μ₁::T = one(T): algorithm parameter in (0, +∞).

σ::T = T(10): algorithm parameter in (1, +∞).

max_eval::Int = -1: maximum number of objective function evaluations.

max_time::Float64 = 30.0: maximum time limit in seconds.

max_iter::Int = typemax(Int): maximum number of iterations.

max_cgiter::Int = 50: subproblem iteration limit.

cgtol::T = T(0.1): subproblem tolerance.

atol::T = √eps(T): absolute tolerance.

rtol::T = √eps(T): relative tolerance, the algorithm stops when ‖∇f(xᵏ)‖ ≤ atol + rtol * ‖∇f(x⁰)‖.

Fatol::T = √eps(T): absolute tolerance on the residual.

Frtol::T = eps(T): relative tolerance on the residual, the algorithm stops when ‖F(xᵏ)‖ ≤ Fatol + Frtol * ‖F(x⁰)‖.

verbose::Int = 0: if > 0, display iteration details every verbose iteration.

The keyword arguments of TronSolverNLS are passed to the TRONTrustRegion constructor.

Output

The value returned is a GenericExecutionStats, see SolverCore.jl.

Callback

The callback is called at each iteration. The expected signature of the callback is callback(nlp, solver, stats), and its output is ignored. Changing any of the input arguments will affect the subsequent iterations. In particular, setting stats.status = :user will stop the algorithm. All relevant information should be available in nlp and solver. Notably, you can access, and modify, the following:

solver.x: current iterate;

solver.gx: current gradient;

stats: structure holding the output of the algorithm (GenericExecutionStats), which contains, among other things:

stats.dual_feas: norm of current gradient;

stats.iter: current iteration counter;

stats.objective: current objective function value;

stats.status: current status of the algorithm. Should be :unknown unless the algorithm attained a stopping criterion. Changing this to anything will stop the algorithm, but you should use :user to properly indicate the intention.

stats.elapsed_time: elapsed time in seconds.

References

This is an adaptation for bound-constrained nonlinear least-squares problems of the TRON method described in

Chih-Jen Lin and Jorge J. Moré, *Newton's Method for Large Bound-Constrained
Optimization Problems*, SIAM J. Optim., 9(4), 1100–1127, 1999.
DOI: 10.1137/S1052623498345075

Examples

using JSOSolvers, ADNLPModels
F(x) = [x[1] - 1.0; 10 * (x[2] - x[1]^2)]
x0 = [-1.2; 1.0]
nls = ADNLSModel(F, x0, 2, zeros(2), 0.5 * ones(2))
stats = tron(nls)

using JSOSolvers, ADNLPModels
F(x) = [x[1] - 1.0; 10 * (x[2] - x[1]^2)]
x0 = [-1.2; 1.0]
nls = ADNLSModel(F, x0, 2, zeros(2), 0.5 * ones(2))
solver = TronSolverNLS(nls)
stats = solve!(solver, nls)

JSOSolvers.tron — Method

tron(nlp; kwargs...)

A pure Julia implementation of a trust-region solver for bound-constrained optimization:

    min f(x)    s.t.    ℓ ≦ x ≦ u

For advanced usage, first define a TronSolver to preallocate the memory used in the algorithm, and then call solve!:

solver = TronSolver(nlp; kwargs...)
solve!(solver, nlp; kwargs...)

Arguments

nlp::AbstractNLPModel{T, V} represents the model to solve, see NLPModels.jl.

The keyword arguments may include

x::V = nlp.meta.x0: the initial guess.

μ₀::T = T(1e-2): algorithm parameter in (0, 0.5).

μ₁::T = one(T): algorithm parameter in (0, +∞).

σ::T = T(10): algorithm parameter in (1, +∞).

max_eval::Int = -1: maximum number of objective function evaluations.

max_time::Float64 = 30.0: maximum time limit in seconds.

max_iter::Int = typemax(Int): maximum number of iterations.

max_cgiter::Int = 50: subproblem's iteration limit.

use_only_objgrad::Bool = false: If true, the algorithm uses only the function objgrad instead of obj and grad.

cgtol::T = T(0.1): subproblem tolerance.

atol::T = √eps(T): absolute tolerance.

rtol::T = √eps(T): relative tolerance, the algorithm stops when ‖∇f(xᵏ)‖ ≤ atol + rtol * ‖∇f(x⁰)‖.

verbose::Int = 0: if > 0, display iteration details every verbose iteration.

The keyword arguments of TronSolver are passed to the TRONTrustRegion constructor.

Output

The value returned is a GenericExecutionStats, see SolverCore.jl.

Callback

The callback is called at each iteration. The expected signature of the callback is callback(nlp, solver, stats), and its output is ignored. Changing any of the input arguments will affect the subsequent iterations. In particular, setting stats.status = :user will stop the algorithm. All relevant information should be available in nlp and solver. Notably, you can access, and modify, the following:

solver.x: current iterate;

solver.gx: current gradient;

stats: structure holding the output of the algorithm (GenericExecutionStats), which contains, among other things:

stats.dual_feas: norm of current gradient;

stats.iter: current iteration counter;

stats.objective: current objective function value;

stats.status: current status of the algorithm. Should be :unknown unless the algorithm has attained a stopping criterion. Changing this to anything will stop the algorithm, but you should use :user to properly indicate the intention.

stats.elapsed_time: elapsed time in seconds.

References

TRON is described in

Chih-Jen Lin and Jorge J. Moré, *Newton's Method for Large Bound-Constrained
Optimization Problems*, SIAM J. Optim., 9(4), 1100–1127, 1999.
DOI: 10.1137/S1052623498345075

Examples

using JSOSolvers, ADNLPModels
nlp = ADNLPModel(x -> sum(x), ones(3), zeros(3), 2 * ones(3));
stats = tron(nlp)

using JSOSolvers, ADNLPModels
nlp = ADNLPModel(x -> sum(x), ones(3), zeros(3), 2 * ones(3));
solver = TronSolver(nlp);
stats = solve!(solver, nlp)

JSOSolvers.trunk — Method

trunk(nls; kwargs...)

A pure Julia implementation of a trust-region solver for nonlinear least-squares problems:

min ½‖F(x)‖²

For advanced usage, first define a TrunkSolverNLS to preallocate the memory used in the algorithm, and then call solve!:

solver = TrunkSolverNLS(nls, subsolver_type::Type{<:KrylovSolver} = LsmrSolver)
solve!(solver, nls; kwargs...)

Arguments

nls::AbstractNLSModel{T, V} represents the model to solve, see NLPModels.jl.

The keyword arguments may include

x::V = nlp.meta.x0: the initial guess.

atol::T = √eps(T): absolute tolerance.

rtol::T = √eps(T): relative tolerance, the algorithm stops when ‖∇f(xᵏ)‖ ≤ atol + rtol * ‖∇f(x⁰)‖.

Fatol::T = √eps(T): absolute tolerance on the residual.

Frtol::T = eps(T): relative tolerance on the residual, the algorithm stops when ‖F(xᵏ)‖ ≤ Fatol + Frtol * ‖F(x⁰)‖.

max_eval::Int = -1: maximum number of objective function evaluations.

max_time::Float64 = 30.0: maximum time limit in seconds.

max_iter::Int = typemax(Int): maximum number of iterations.

bk_max::Int = 10: algorithm parameter.

monotone::Bool = true: algorithm parameter.

nm_itmax::Int = 25: algorithm parameter.

verbose::Int = 0: if > 0, display iteration details every verbose iteration.

verbose_subsolver::Int = 0: if > 0, display iteration information every verbose_subsolver iteration of the subsolver.

See JSOSolvers.trunkls_allowed_subsolvers for a list of available KrylovSolver.

Output

The value returned is a GenericExecutionStats, see SolverCore.jl.

Callback

The callback is called at each iteration. The expected signature of the callback is callback(nlp, solver, stats), and its output is ignored. Changing any of the input arguments will affect the subsequent iterations. In particular, setting stats.status = :user will stop the algorithm. All relevant information should be available in nlp and solver. Notably, you can access, and modify, the following:

solver.x: current iterate;

solver.gx: current gradient;

stats: structure holding the output of the algorithm (GenericExecutionStats), which contains, among other things:

stats.dual_feas: norm of current gradient;

stats.iter: current iteration counter;

stats.objective: current objective function value;

stats.status: current status of the algorithm. Should be :unknown unless the algorithm attained a stopping criterion. Changing this to anything will stop the algorithm, but you should use :user to properly indicate the intention.

stats.elapsed_time: elapsed time in seconds.

References

This implementation follows the description given in

A. R. Conn, N. I. M. Gould, and Ph. L. Toint,
Trust-Region Methods, volume 1 of MPS/SIAM Series on Optimization.
SIAM, Philadelphia, USA, 2000.
DOI: 10.1137/1.9780898719857

The main algorithm follows the basic trust-region method described in Section 6. The backtracking linesearch follows Section 10.3.2. The nonmonotone strategy follows Section 10.1.3, Algorithm 10.1.2.

Examples

using JSOSolvers, ADNLPModels
F(x) = [x[1] - 1.0; 10 * (x[2] - x[1]^2)]
x0 = [-1.2; 1.0]
nls = ADNLSModel(F, x0, 2)
stats = trunk(nls)

using JSOSolvers, ADNLPModels
F(x) = [x[1] - 1.0; 10 * (x[2] - x[1]^2)]
x0 = [-1.2; 1.0]
nls = ADNLSModel(F, x0, 2)
solver = TrunkSolverNLS(nls)
stats = solve!(solver, nls)

JSOSolvers.trunk — Method

trunk(nlp; kwargs...)

A trust-region solver for unconstrained optimization using exact second derivatives.

For advanced usage, first define a TrunkSolver to preallocate the memory used in the algorithm, and then call solve!:

solver = TrunkSolver(nlp, subsolver_type::Type{<:KrylovSolver} = CgSolver)
solve!(solver, nlp; kwargs...)

Arguments

nlp::AbstractNLPModel{T, V} represents the model to solve, see NLPModels.jl.

The keyword arguments may include

subsolver_logger::AbstractLogger = NullLogger(): subproblem's logger.

x::V = nlp.meta.x0: the initial guess.

atol::T = √eps(T): absolute tolerance.

rtol::T = √eps(T): relative tolerance, the algorithm stops when ‖∇f(xᵏ)‖ ≤ atol + rtol * ‖∇f(x⁰)‖.

max_eval::Int = -1: maximum number of objective function evaluations.

max_time::Float64 = 30.0: maximum time limit in seconds.

max_iter::Int = typemax(Int): maximum number of iterations.

bk_max::Int = 10: algorithm parameter.

monotone::Bool = true: algorithm parameter.

nm_itmax::Int = 25: algorithm parameter.

verbose::Int = 0: if > 0, display iteration information every verbose iteration.

verbose_subsolver::Int = 0: if > 0, display iteration information every verbose_subsolver iteration of the subsolver.

Output

The returned value is a GenericExecutionStats, see SolverCore.jl.

Callback

The callback is called at each iteration. The expected signature of the callback is callback(nlp, solver, stats), and its output is ignored. Changing any of the input arguments will affect the subsequent iterations. In particular, setting stats.status = :user will stop the algorithm. All relevant information should be available in nlp and solver. Notably, you can access, and modify, the following:

solver.x: current iterate;

solver.gx: current gradient;

stats: structure holding the output of the algorithm (GenericExecutionStats), which contains, among other things:

stats.dual_feas: norm of current gradient;

stats.iter: current iteration counter;

stats.objective: current objective function value;

stats.status: current status of the algorithm. Should be :unknown unless the algorithm attained a stopping criterion. Changing this to anything will stop the algorithm, but you should use :user to properly indicate the intention.

stats.elapsed_time: elapsed time in seconds.

References

This implementation follows the description given in

A. R. Conn, N. I. M. Gould, and Ph. L. Toint,
Trust-Region Methods, volume 1 of MPS/SIAM Series on Optimization.
SIAM, Philadelphia, USA, 2000.
DOI: 10.1137/1.9780898719857

The main algorithm follows the basic trust-region method described in Section 6. The backtracking linesearch follows Section 10.3.2. The nonmonotone strategy follows Section 10.1.3, Algorithm 10.1.2.

Examples

using JSOSolvers, ADNLPModels
nlp = ADNLPModel(x -> sum(x.^2), ones(3))
stats = trunk(nlp)

using JSOSolvers, ADNLPModels
nlp = ADNLPModel(x -> sum(x.^2), ones(3))
solver = TrunkSolver(nlp)
stats = solve!(solver, nlp)