Solvers

Solver list

lbfgs

tron

trunk

Problem type	Solvers

Unconstrained NLP	lbfgs, tron, trunk

Unconstrained NLS	trunk, tron

Bound-constrained NLP	tron

Bound-constrained NLS	tron

Solver list

JSOSolvers.lbfgs — Function

lbfgs(nlp; kwargs...)

An implementation of a limited memory BFGS line-search method for unconstrained minimization.

For advanced usage, first define a LBFGSSolver to preallocate the memory used in the algorithm, and then call solve!.

solver = LBFGSSolver(nlp; mem::Int = 5)
solve!(solver, nlp; kwargs...)

Arguments

nlp::AbstractNLPModel{T, V} represents the model to solve, see NLPModels.jl.

The keyword arguments may include

x::V = nlp.meta.x0: the initial guess.

atol::T = √eps(T): absolute tolerance.

rtol::T = √eps(T): relative tolerance, the algorithm stops when ||∇f(xᵏ)|| ≤ atol + rtol * ||∇f(x⁰)||.

max_eval::Int = -1: maximum number of objective function evaluations.

max_time::Float64 = 30.0: maximum time limit in seconds.

verbose::Int = 0: if > 0, display iteration details every verbose iteration.

mem::Int = 5: memory parameter of the lbfgs algorithm.

Output

The returned value is a GenericExecutionStats, see SolverCore.jl.

Examples

using JSOSolvers, ADNLPModels
nlp = ADNLPModel(x -> sum(x.^2), ones(3));
stats = lbfgs(nlp)

# output

"Execution stats: first-order stationary"

using JSOSolvers, ADNLPModels
nlp = ADNLPModel(x -> sum(x.^2), ones(3));
solver = LBFGSSolver(nlp; mem = 5);
stats = solve!(solver, nlp)

# output

"Execution stats: first-order stationary"

JSOSolvers.tron — Function

tron(nlp; kwargs...)

A pure Julia implementation of a trust-region solver for bound-constrained optimization:

    min f(x)    s.t.    ℓ ≦ x ≦ u

For advanced usage, first define a TronSolver to preallocate the memory used in the algorithm, and then call solve!.

solver = TronSolver(nlp)
solve!(solver, nlp; kwargs...)

Arguments

nlp::AbstractNLPModel{T, V} represents the model to solve, see NLPModels.jl.

The keyword arguments may include

subsolver_logger::AbstractLogger = NullLogger(): subproblem's logger.

x::V = nlp.meta.x0: the initial guess.

μ₀::T = T(1e-2): algorithm parameter in (0, 0.5).

μ₁::T = one(T): algorithm parameter in (0, +∞).

σ::T = T(10): algorithm parameter in (1, +∞).

max_eval::Int = -1: maximum number of objective function evaluations.

max_time::Float64 = 30.0: maximum time limit in seconds.

max_cgiter::Int = 50: subproblem's iteration limit.

use_only_objgrad::Bool = false: If true, the algorithm uses only the function objgrad instead of obj and grad.

cgtol::T = T(0.1): subproblem tolerance.

atol::T = √eps(T): absolute tolerance.

rtol::T = √eps(T): relative tolerance, the algorithm stops when ||∇f(xᵏ)|| ≤ atol + rtol * ||∇f(x⁰)||.

verbose::Int = 0: if > 0, display iteration details every verbose iteration.

Output

The value returned is a GenericExecutionStats, see SolverCore.jl.

References

TRON is described in

Chih-Jen Lin and Jorge J. Moré, *Newton's Method for Large Bound-Constrained
Optimization Problems*, SIAM J. Optim., 9(4), 1100–1127, 1999.
DOI: 10.1137/S1052623498345075

Examples

using JSOSolvers, ADNLPModels
nlp = ADNLPModel(x -> sum(x), ones(3), zeros(3), 2 * ones(3));
stats = tron(nlp)

using JSOSolvers, ADNLPModels
nlp = ADNLPModel(x -> sum(x), ones(3), zeros(3), 2 * ones(3));
solver = TronSolver(nlp);
stats = solve!(solver, nlp)

tron(nls; kwargs...)

A pure Julia implementation of a trust-region solver for bound-constrained nonlinear least-squares problems:

min ½‖F(x)‖²    s.t.    ℓ ≦ x ≦ u

Arguments

nls::AbstractNLSModel{T, V} represents the model to solve, see NLPModels.jl.

The keyword arguments may include

subsolver_logger::AbstractLogger = NullLogger(): subproblem's logger.

x::V = nlp.meta.x0: the initial guess.

subsolver::Symbol = :lsmr: Krylov.jl method used as subproblem solver, see JSOSolvers.tronls_allowed_subsolvers for a list.

μ₀::T = T(1e-2): algorithm parameter in (0, 0.5).

μ₁::T = one(T): algorithm parameter in (0, +∞).

σ::T = T(10): algorithm parameter in (1, +∞).

max_eval::Int = -1: maximum number of objective function evaluations.

max_time::Float64 = 30.0: maximum time limit in seconds.

max_cgiter::Int = 50: subproblem iteration limit.

cgtol::T = T(0.1): subproblem tolerance.

atol::T = √eps(T): absolute tolerance.

rtol::T = √eps(T): relative tolerance, the algorithm stops when ||∇f(xᵏ)|| ≤ atol + rtol * ||∇f(x⁰)||.

verbose::Int = 0: if > 0, display iteration details every verbose iteration.

Output

The value returned is a GenericExecutionStats, see SolverCore.jl.

References

This is an adaptation for bound-constrained nonlinear least-squares problems of the TRON method described in

Chih-Jen Lin and Jorge J. Moré, *Newton's Method for Large Bound-Constrained
Optimization Problems*, SIAM J. Optim., 9(4), 1100–1127, 1999.
DOI: 10.1137/S1052623498345075

Examples

using JSOSolvers, ADNLPModels
F(x) = [x[1] - 1.0; 10 * (x[2] - x[1]^2)]
x0 = [-1.2; 1.0]
nls = ADNLSModel(F, x0, 2, zeros(2), 0.5 * ones(2))
stats = tron(nls)

JSOSolvers.trunk — Function

trunk(nlp; kwargs...)

A trust-region solver for unconstrained optimization using exact second derivatives.

For advanced usage, first define a TrunkSolver to preallocate the memory used in the algorithm, and then call solve!.

solver = TrunkSolver(nlp, subsolver_type::Type{<:KrylovSolver} = CgSolver)
solve!(solver, nlp; kwargs...)

Arguments

nlp::AbstractNLPModel{T, V} represents the model to solve, see NLPModels.jl.

The keyword arguments may include

subsolver_logger::AbstractLogger = NullLogger(): subproblem's logger.

x::V = nlp.meta.x0: the initial guess.

atol::T = √eps(T): absolute tolerance.

rtol::T = √eps(T): relative tolerance, the algorithm stops when ||∇f(xᵏ)|| ≤ atol + rtol * ||∇f(x⁰)||.

max_eval::Int = -1: maximum number of objective function evaluations.

max_time::Float64 = 30.0: maximum time limit in seconds.

bk_max::Int = 10: algorithm parameter.

monotone::Bool = true: algorithm parameter.

nm_itmax::Int = 25: algorithm parameter.

verbose::Int = 0: if > 0, display iteration information every verbose iteration.

verbose_subsolver::Int = 0: if > 0, display iteration information every verbose_subsolver iteration of the subsolver.

Output

The returned value is a GenericExecutionStats, see SolverCore.jl.

References

This implementation follows the description given in

A. R. Conn, N. I. M. Gould, and Ph. L. Toint,
Trust-Region Methods, volume 1 of MPS/SIAM Series on Optimization.
SIAM, Philadelphia, USA, 2000.
DOI: 10.1137/1.9780898719857

The main algorithm follows the basic trust-region method described in Section 6. The backtracking linesearch follows Section 10.3.2. The nonmonotone strategy follows Section 10.1.3, Algorithm 10.1.2.

Examples

using JSOSolvers, ADNLPModels
nlp = ADNLPModel(x -> sum(x.^2), ones(3))
stats = trunk(nlp)

using JSOSolvers, ADNLPModels
nlp = ADNLPModel(x -> sum(x.^2), ones(3))
solver = TrunkSolver(nlp)
stats = solve!(solver, nlp)

trunk(nls; kwargs...)

A pure Julia implementation of a trust-region solver for nonlinear least-squares problems:

min ½‖F(x)‖²

Arguments

nls::AbstractNLSModel{T, V} represents the model to solve, see NLPModels.jl.

The keyword arguments may include

x::V = nlp.meta.x0: the initial guess.

subsolver::Symbol = :lsmr: Krylov.jl method used as subproblem solver, see JSOSolvers.trunkls_allowed_subsolvers for a list.

atol::T = √eps(T): absolute tolerance.

rtol::T = √eps(T): relative tolerance, the algorithm stops when ||∇f(xᵏ)|| ≤ atol + rtol * ||∇f(x⁰)||.

max_eval::Int = -1: maximum number of objective function evaluations.

max_time::Float64 = 30.0: maximum time limit in seconds.

bk_max::Int = 10: algorithm parameter.

monotone::Bool = true: algorithm parameter.

nm_itmax::Int = 25: algorithm parameter.

trsolver_args::Dict{Symbol, Any} = Dict{Symbol, Any}(): additional keyword arguments for the subproblem solver.

verbose::Int = 0: if > 0, display iteration details every verbose iteration.

Output

The value returned is a GenericExecutionStats, see SolverCore.jl.

References

This implementation follows the description given in

A. R. Conn, N. I. M. Gould, and Ph. L. Toint,
Trust-Region Methods, volume 1 of MPS/SIAM Series on Optimization.
SIAM, Philadelphia, USA, 2000.
DOI: 10.1137/1.9780898719857

The main algorithm follows the basic trust-region method described in Section 6. The backtracking linesearch follows Section 10.3.2. The nonmonotone strategy follows Section 10.1.3, Algorithm 10.1.2.

Examples

using JSOSolvers, ADNLPModels
F(x) = [x[1] - 1.0; 10 * (x[2] - x[1]^2)]
x0 = [-1.2; 1.0]
nls = ADNLSModel(F, x0, 2)
stats = trunk(nls)